PEP8 renames Settings and Parameters modules
