+++ _RunOffline.py
@@ -24,7 +24,7 @@ if os.getcwd() !=  _topMouseTrackerDir : #If the current working directory is no
 
 # Loads TMT modules#
 
-import TopMouseTracker.Parameters as params
+import TopMouseTracker.parameters as params
 import TopMouseTracker.utilities as utils
 import TopMouseTracker.IO as IO
 import TopMouseTracker._Tracker as tracker
